Split qvalue.Equals out of qvalue.go
